# contracts/openapi.yaml
...
openapi: 3.1.0
...
info:
  title: cb API
  version: 1.0.0
  description: |-
    REST API of cb-backend, consumed by cb-dashboard (session cookie) and cb-env (device bearer token).

    Every success response is `{ success: true, data, meta: { correlationId } }`; paginated lists add
    `meta.pagination`. Every error is `ErrorBody` (PRD §12.7). Stored secrets are write-only: credential
    fields appear only in request bodies and are never returned. The `/tunnel` WebSocket is specified in
    `tunnel-protocol.md`.
servers:
  - url: http://localhost:4200
    description: Local development
...
paths:
...
  /api/resources/{resourceId}:
    patch:
      operationId: updateResource
      tags:
        - Resources
      summary: Update or rotate a resource (admin)
      security:
        - sessionCookie: []
        - deviceToken: []
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResourceBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Resource"
                    required:
                      - data
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteResource
      tags:
        - Resources
      summary: Delete a resource with its variables and profiles (admin)
      security:
        - sessionCookie: []
        - deviceToken: []
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deleted:
                            type: boolean
                            const: true
                        required:
                          - deleted
                    required:
                      - data
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/resources/{resourceId}/test:
    post:
      operationId: testResource
      tags:
        - Resources
      summary: Test a resource's real credentials from the gateway (admin)
      description: Connects to the real service with the credentials of `profile` (default "default") and reports the outcome. The credential is never returned, and is scrubbed from error messages.
      security:
        - sessionCookie: []
        - deviceToken: []
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                profile:
                  type: string
                  description: Credential profile to test.
                  default: default
      responses:
        "200":
          description: Test finished (check `ok`)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ResourceTestResult"
                    required:
                      - data
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/resources/{resourceId}/profiles:
    get:
      operationId: listProfiles
      tags:
        - Resources
      summary: Credential profiles of a resource ("default" first)
      security:
        - sessionCookie: []
        - deviceToken: []
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Profile"
                    required:
                      - data
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createProfile
      tags:
        - Resources
      summary: Create a credential profile (admin); credentials are write-only
      security:
        - sessionCookie: []
        - deviceToken: []
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Profile"
                    required:
                      - data
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/resources/{resourceId}/profiles/{name}:
    put:
      operationId: rotateProfile
      tags:
        - Resources
      summary: Rotate a credential profile (admin)
      description: Takes effect on the next connection; live tunnels keep their session.
      security:
        - sessionCookie: []
        - deviceToken: []
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/profileName"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileCredentialsBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Profile"
                    required:
                      - data
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteProfile
      tags:
        - Resources
      summary: Delete a credential profile (admin; 409 while assigned in a grant)
      security:
        - sessionCookie: []
        - deviceToken: []
      parameters:
        - $ref: "#/components/parameters/resourceId"
        - $ref: "#/components/parameters/profileName"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deleted:
                            type: boolean
                            const: true
                        required:
                          - deleted
                    required:
                      - data
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/variables/{variableId}:
    patch:
      operationId: updateVariable
      tags:
        - Variables
      summary: Update a variable (admin)
      security:
        - sessionCookie: []
        - deviceToken: []
      parameters:
        - $ref: "#/components/parameters/variableId"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVariableBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Variable"
                    required:
                      - data
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteVariable
      tags:
        - Variables
      summary: Delete a variable (admin)
      security:
        - sessionCookie: []
        - deviceToken: []
      parameters:
        - $ref: "#/components/parameters/variableId"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deleted:
                            type: boolean
                            const: true
                        required:
                          - deleted
                    required:
                      - data
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/grants/{grantId}:
    patch:
      operationId: updateGrant
      tags:
        - Access
      summary: Change a grant's expiry or credential profiles (admin)
      security:
        - sessionCookie: []
        - deviceToken: []
      parameters:
        - $ref: "#/components/parameters/grantId"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGrantBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Grant"
                    required:
                      - data
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: revokeGrant
      tags:
        - Access
      summary: Revoke a grant now; live tunnels close (FR-GW-007)
      security:
        - sessionCookie: []
        - deviceToken: []
      parameters:
        - $ref: "#/components/parameters/grantId"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          revoked:
                            type: boolean
                            const: true
                        required:
                          - revoked
                    required:
                      - data
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cb_session
      description: "Dashboard session (httpOnly, SameSite=Lax). Mutating requests must also send `x-cb-csrf: 1` (see the CsrfHeader parameter)."
    deviceToken:
      type: http
      scheme: bearer
      description: CLI device token (prefix `cbd_`) from the device-code flow. Used by `/api/cli/*` and `/api/agent/*`.
  parameters:
    CsrfHeader:
      name: x-cb-csrf
      in: header
      required: true
      description: Must be `1` on every mutating request authenticated with the session cookie (otherwise 403 CSRF_REQUIRED). Ignored for device bearer tokens.
      schema:
        type: string
        const: "1"
    CorrelationId:
      name: x-correlation-id
      in: header
      required: false
      description: Optional; generated when absent and echoed in the response header and meta.
      schema:
        type: string
...
    resourceId:
      name: resourceId
      in: path
      required: true
      description: Resource id
      schema:
        $ref: "#/components/schemas/ObjectId"
    variableId:
      name: variableId
      in: path
      required: true
      description: Variable id
      schema:
        $ref: "#/components/schemas/ObjectId"
    grantId:
      name: grantId
      in: path
      required: true
      description: Grant id
      schema:
        $ref: "#/components/schemas/ObjectId"
...
    profileName:
      name: name
      in: path
      required: true
      description: Credential profile name.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: VALIDATION_FAILED (or DEVICE_AUTH_PENDING on the device poll)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            success: false
            error:
              code: VALIDATION_FAILED
              message: The request is invalid.
              statusCode: 400
              details:
                - path: name
                  message: "Too small: expected string to have >=1 characters"
              correlationId: c0ffee12-0000-4000-8000-000000000000
    Unauthorized:
      description: UNAUTHORIZED or INVALID_CREDENTIALS
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication is required.
              statusCode: 401
              correlationId: c0ffee12-0000-4000-8000-000000000000
    Forbidden:
      description: FORBIDDEN, CSRF_REQUIRED, NO_ACCESS or ENVIRONMENT_KILLED
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            success: false
            error:
              code: FORBIDDEN
              message: You do not have access to this resource.
              statusCode: 403
              correlationId: c0ffee12-0000-4000-8000-000000000000
    NotFound:
      description: NOT_FOUND (malformed ids are also reported as not found)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            success: false
            error:
              code: NOT_FOUND
              message: The requested resource was not found.
              statusCode: 404
              correlationId: c0ffee12-0000-4000-8000-000000000000
    Conflict:
      description: CONFLICT, EMAIL_TAKEN or LAST_OWNER
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            success: false
            error:
              code: CONFLICT
              message: The resource already exists or conflicts with another.
              statusCode: 409
              correlationId: c0ffee12-0000-4000-8000-000000000000
...
  schemas:
    ResourceTestResult:
      type: object
      properties:
        ok:
          type: boolean
        profile:
          type: string
        latencyMs:
          type: integer
        message:
          type: string
          description: Outcome in plain words; never contains the credential.
      required: [ok, profile, latencyMs, message]
...
    ObjectId:
      type: string
      pattern: ^[a-f0-9]{24}$
      examples:
        - 665f1c2ab3e4d5f6a7b8c9d0
...
    ResourceKind:
      type: string
      enum:
        - mongodb
        - redis
        - postgres
        - mysql
        - smtp
        - http
        - oauth
        - aws
        - google-sa
        - apns
    VariableType:
      type: string
      enum:
        - plain
        - generated
        - visible
        - brokered
    GeneratedFormat:
      type: string
      enum:
        - hex:32
        - hex:64
        - base64:32
        - base64url:32
        - alnum:32
        - alnum:48
        - uuid
    AuthScheme:
      type: string
      enum:
        - bearer
        - x-api-key
        - basic-password
...
    Meta:
      type: object
      properties:
        correlationId:
          type: string
      required:
        - correlationId
...
    Envelope:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data: {}
        meta:
          $ref: "#/components/schemas/Meta"
      required:
        - success
        - data
        - meta
...
    ErrorDetail:
      type: object
      properties:
        path:
          type: string
        message:
          type: string
      required:
        - path
        - message
    ErrorBody:
      type: object
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          properties:
            code:
              type: string
              description: SCREAMING_SNAKE_CASE code from the error catalog.
              examples:
                - NOT_FOUND
            message:
              type: string
            statusCode:
              type: integer
            details:
              type: array
              items:
                $ref: "#/components/schemas/ErrorDetail"
            correlationId:
              type: string
          required:
            - code
            - message
            - statusCode
            - correlationId
      required:
        - success
        - error
...
    Resource:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        environmentId:
          $ref: "#/components/schemas/ObjectId"
        kind:
          $ref: "#/components/schemas/ResourceKind"
        name:
          type: string
        config:
          type: object
          additionalProperties: true
          description: Non-secret configuration (e.g. host, database, tls, upstreamUrl, authScheme, fakePrefix, basePath, redirectHosts, region, endpoint, tokenUrl, projectId, clientEmail, keyId, teamId). Never contains credentials.
        credentialsSet:
          type: boolean
        rotatedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        disabled:
          type: boolean
        brokeredFields:
          type: array
          items:
            type: string
          description: Variable fields this resource can broker.
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - environmentId
        - kind
        - name
        - config
        - credentialsSet
        - rotatedAt
        - disabled
        - brokeredFields
        - createdAt
      description: Credentials are write-only and never returned (L15, S2).
    Profile:
      type: object
      properties:
        name:
          type: string
        rotatedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        isDefault:
          type: boolean
      required:
        - name
        - rotatedAt
        - isDefault
    Variable:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        environmentId:
          $ref: "#/components/schemas/ObjectId"
        key:
          type: string
        type:
          $ref: "#/components/schemas/VariableType"
        required:
          type: boolean
        value:
          oneOf:
            - type: string
              description: Only plain values are readable; null for every other type.
            - type: "null"
        format:
          oneOf:
            - type: string
            - type: "null"
        resourceId:
          oneOf:
            - $ref: "#/components/schemas/ObjectId"
            - type: "null"
        resourceName:
          oneOf:
            - type: string
            - type: "null"
        field:
          oneOf:
            - type: string
            - type: "null"
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - environmentId
        - key
        - type
        - required
        - value
        - format
        - resourceId
        - resourceName
        - field
        - updatedAt
...
    ResourceProfileAssignment:
      type: object
      properties:
        resourceId:
          $ref: "#/components/schemas/ObjectId"
        profile:
          type: string
          minLength: 1
          maxLength: 30
      required:
        - resourceId
        - profile
    Grant:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        scope:
          type: string
          enum:
            - environment
            - project
        projectId:
          $ref: "#/components/schemas/ObjectId"
        environmentId:
          oneOf:
            - $ref: "#/components/schemas/ObjectId"
            - type: "null"
        userId:
          $ref: "#/components/schemas/ObjectId"
        resourceProfiles:
          type: array
          items:
            $ref: "#/components/schemas/ResourceProfileAssignment"
          description: Resources not listed use the "default" profile.
        expiresAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        createdAt:
          type: string
          format: date-time
        createdBy:
          $ref: "#/components/schemas/ObjectId"
      required:
        - id
        - scope
        - projectId
        - environmentId
        - userId
        - resourceProfiles
        - expiresAt
        - createdAt
        - createdBy
...
    UpdateResourceBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
          description: Trimmed.
        disabled:
          type: boolean
        connectionUri:
          type: string
          minLength: 8
          writeOnly: true
          description: Rotates the credential (database/SMTP kinds).
        apiKey:
          type: string
          minLength: 1
          writeOnly: true
          description: Rotates the credential (http).
        clientSecret:
          type: string
          minLength: 1
          writeOnly: true
          description: Rotates the credential (oauth).
        accessKeyId:
          type: string
          minLength: 3
          writeOnly: true
          description: Rotates the credential with secretAccessKey (aws).
        secretAccessKey:
          type: string
          minLength: 8
          writeOnly: true
          description: Rotates the credential with accessKeyId (aws).
        serviceAccountJson:
          type: string
          description: Rotates the credential (google-sa).
          writeOnly: true
        privateKey:
          type: string
          description: Rotates the credential (apns).
          writeOnly: true
        keyId:
          type: string
          pattern: ^[A-Z0-9]{10}$
        teamId:
          type: string
          pattern: ^[A-Z0-9]{10}$
        region:
          type: string
          pattern: ^[a-z0-9-]+$
        endpoint:
          type: string
          format: uri
        tokenUrl:
          type: string
          format: uri
        upstreamUrl:
          type: string
          format: uri
        authScheme:
          $ref: "#/components/schemas/AuthScheme"
        fakePrefix:
          type: string
          maxLength: 20
        basePath:
          type: string
        redirectHosts:
          type: array
          items:
            type: string
            pattern: ^[A-Za-z0-9.-]+:[0-9]{1,5}$
            description: host:port, e.g. api.example.com:443 (case-insensitive; stored lowercased).
            examples:
              - api.example.com:443
          maxItems: 10
      description: All fields optional; fields that do not apply to the resource's kind are ignored.
    ProfileCredentialsBody:
      type: object
      properties:
        connectionUri:
          type: string
          minLength: 8
          writeOnly: true
          description: Database/SMTP kinds.
        apiKey:
          type: string
          minLength: 1
          writeOnly: true
          description: http.
        clientSecret:
          type: string
          minLength: 1
          writeOnly: true
          description: oauth.
        accessKeyId:
          type: string
          minLength: 3
          writeOnly: true
          description: aws (with secretAccessKey).
        secretAccessKey:
          type: string
          minLength: 8
          writeOnly: true
          description: aws (with accessKeyId).
        serviceAccountJson:
          type: string
          description: google-sa.
          writeOnly: true
        privateKey:
          type: string
          description: apns.
          writeOnly: true
      description: The write-only credential fields matching the resource's kind; one set per profile.
    CreateProfileBody:
      allOf:
        - $ref: "#/components/schemas/ProfileCredentialsBody"
        - type: object
          properties:
            name:
              type: string
              pattern: ^[a-z][a-z0-9-]{0,29}$
              not:
                const: default
              description: Lowercase letters, digits and dashes, e.g. readonly. "default" is reserved.
          required:
            - name
...
    VariableKey:
      type: string
      pattern: ^[A-Z_][A-Z0-9_]*$
      description: UPPER_SNAKE_CASE (trimmed).
      examples:
        - DATABASE_URL
...
    UpdateVariableBody:
      type: object
      properties:
        key:
          $ref: "#/components/schemas/VariableKey"
        value:
          type: string
          maxLength: 10000
          writeOnly: true
          description: "plain: new value; visible: new write-only value."
        format:
          $ref: "#/components/schemas/GeneratedFormat"
        resourceId:
          type: string
        field:
          type: string
        required:
          type: boolean
      description: Fields that do not apply to the variable's type are ignored.
    ExpiresAt:
      oneOf:
        - type: string
          format: date-time
          description: ISO 8601; must be in the future. null = permanent.
        - type: "null"
    ResourceProfiles:
      type: array
      items:
        $ref: "#/components/schemas/ResourceProfileAssignment"
      maxItems: 50
...
    UpdateGrantBody:
      type: object
      properties:
        resourceProfiles:
          $ref: "#/components/schemas/ResourceProfiles"
        expiresAt:
          $ref: "#/components/schemas/ExpiresAt"
...